requirements.txt
fastapi
uvicorn[standard]
python-dotenv
transformers
sentencepiece
texttable
toml
aioredis
pyyaml